+++ changelog.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## Breaking Change
+
+### Changed
+
+- ON-4232 # Forms.generateFormUrl() now takes an object with the following properties:
+  - formId
+  - externalId
+  - preFillData
+  - expiryInSeconds
+
 ## 0.2.9 (2019-06-03)
 
 ### Added